Conversational_UI/Releases/NLPEngine/requirements.txt
kafka-python~=2.0.2
rasa~=2.3.1
spacy~=3.0.3
nltk~=3.5
flask~=1.1.2